moved setPlaces logic to separate file
