+++ routes/index.js
@@ -1,41 +1,13 @@
 ﻿var express = require('express');
 var router = express.Router();
-
-function setPlaces(topten) {
-    var place = 1;
-    for (var i = 0; i < topten.length; i++) {
-        if (i == 0) {
-            // First team, so set it to 1st
-            topten[i]["Place"] = Number(place);
-            topten[i]["Ordinal"] = ordinal(place);
-            continue;
-        } else {
-            if (topten[i]["Team_Score"] < topten[i - 1]["Team_Score"]) {
-                // Not tied with previous team
-                topten[i]["Place"] = Number(++place);
-                topten[i]["Ordinal"] = ordinal(place);
-            } else {
-                // Tied with previous team
-                topten[i]["Place"] = Number(place);
-                topten[i]["Ordinal"] = ordinal(place);
-            }
-        }
-    }
-    return topten;
-}
-
-function ordinal(n) {
-	var s = ["th", "st", "nd", "rd"],
-		v = n % 100;
-	return n + (s[(v - 20) % 10] || s[v] || s[0]);
-}
+var places = require('../public/javascripts/set-places.js')
 
 /* Raw JSON page */
 router.get('/raw', function(req, res) {
     var db = req.db;
     var teams = db.get('Teams');
     teams.find({}, { "sort" : [['Team_Score','desc']] }, function (e, docs) {
-        docs = setPlaces(docs);
+        docs = places(docs);
         res.render('raw', {
             "rawjson" : JSON.stringify(docs, null, 2),
         });
@@ -47,7 +19,7 @@ router.get('/', function (req, res) {
     var db = req.db;
     var teams = db.get('Teams');
     teams.find({}, { "sort" : [['Team_Score','desc']], "limit" : 10 }, function (e, docs) {
-        docs = setPlaces(docs);
+        docs = places(docs);
         res.render('index', {
             "teamlist" : docs,
             "event" : "Event Name"
